+++ crates/tantivy/src/columnar/columnar/writer/mod.rs
@@ -468,7 +468,7 @@ impl ColumnarWriter {
     }
 }
 
-// Serialize [Dictionary, Column, dictionary num bytes U32::LE]
+// Serialize [Dictionary, Column, dictionary num bytes U64::LE]
 // Column: [Column Index, Column Values, column index num bytes U32::LE]
 #[allow(clippy::too_many_arguments)]
 fn serialize_bytes_or_str_column(
@@ -489,7 +489,7 @@ fn serialize_bytes_or_str_column(
     let mut counting_writer = CountingWriter::wrap(wrt);
     let term_id_mapping: TermIdMapping =
         dictionary_builder.serialize(arena, &mut counting_writer)?;
-    let dictionary_num_bytes: u32 = counting_writer.written_bytes() as u32;
+    let dictionary_num_bytes: u64 = counting_writer.written_bytes();
     let mut wrt = counting_writer.finish();
     let operation_iterator = operation_it.map(|symbol: ColumnOperation<UnorderedId>| {
         // We map unordered ids to ordered ids.
